+++ src/main/groovy/wslite/http/HTTPMethod.groovy
@@ -1,4 +1,4 @@
-/* Copyright 2011 the original author or authors.
+/* Copyright 2011-2014 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,6 +19,8 @@ enum HTTPMethod {
     GET('GET'),
     POST('POST'),
     PUT('PUT'),
+    PATCH('PATCH'),
+    HEAD('HEAD'),
     DELETE('DELETE')
 
     private String method