+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0 (2014-07-09)
+
+* RESTClient support for HEAD and PATCH methods
+* SOAPClient should allow DOCTYPE in response
+
 ## 1.0.0 (2014-07-09)
 
 ## 0.8.2 (2014-07-09)